# docs/swagger.yaml
basePath: /
definitions:
  article.ArticleModel:
    properties:
      created_at:
        type: string
      filepath:
        type: string
      files:
        items:
          $ref: '#/definitions/article.ArticlesFilesDBModel'
        type: array
      tags:
        type: string
      text:
        type: string
      title:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    required:
    - created_at
    - filepath
    - files
    - tags
    - text
    - title
    - updated_at
    - uuid
    type: object
  article.ArticleSuccessModel:
    properties:
      success:
        type: boolean
    type: object
  article.ArticleUuidModel:
    properties:
      uuid:
        type: string
    required:
    - uuid
    type: object
  article.ArticlesFilesDBModel:
    properties:
      filename:
        type: string
      filepath:
        type: string
      files_id:
        type: integer
      index:
        type: integer
    required:
    - filename
    - filepath
    - index
    type: object
  article.ArticlesModel:
    properties:
      articles:
        items:
          $ref: '#/definitions/article.ArticleModel'
        type: array
    required:
    - articles
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  title: MISU Main Server
  version: "1.0"
paths:
...
  /user/article/create:
    post:
      consumes:
      - application/json
      description: Update article
      operationId: update-article
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            $ref: '#/definitions/article.ArticleSuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: UpdateArticle
      tags:
      - article
  /user/article/delete:
    post:
      consumes:
      - application/json
      description: Удаление статьи
      operationId: delete-article
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/article.ArticleUuidModel'
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            $ref: '#/definitions/article.ArticleSuccessModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: DeleteArticle
      tags:
      - article
  /user/article/get:
    post:
      consumes:
      - application/json
      description: Get information about article
      operationId: get-article
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/article.ArticleUuidModel'
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            $ref: '#/definitions/article.ArticleModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: GetArticle
      tags:
      - article
  /user/article/get/all:
    post:
      consumes:
      - application/json
      description: Получение списка статей
      operationId: get-articles
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            $ref: '#/definitions/article.ArticlesModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: GetArticles
      tags:
      - article
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"